requirements.txt
streamlit>=1.32.0
openai>=1.14.0
psycopg2-binary>=2.9.9
python-dotenv
numpy
scikit-learn
pytest
fastapi
uvicorn
pydantic>=1.4.0
pandas>=2.2.0
pytest>=8.0.0
